+++ extensions/servicenow/src/login-to-selected-instance.ts
@@ -1,5 +1,6 @@
 import { LocalStorage, open, showToast, Toast } from "@raycast/api";
 import { Instance } from "./types";
+import { getInstanceBaseUrl } from "./utils/instanceUrl";
 
 export default async function main() {
   const instance = await LocalStorage.getItem<string>("selected-instance");
@@ -10,6 +11,6 @@ export default async function main() {
 
   const instanceProfile = JSON.parse(instance) as Instance;
   open(
-    `https://${instanceProfile.name}.service-now.com/login.do?user_name=${instanceProfile.username}&user_password=${instanceProfile.password}&sys_action=sysverb_login`,
+    `${getInstanceBaseUrl(instanceProfile)}/login.do?user_name=${instanceProfile.username}&user_password=${instanceProfile.password}&sys_action=sysverb_login`,
   );
 }